Homework2/Component1/CloudFunctions/ExportFeedbackData/requirements.txt
functions-framework==3.*
google-cloud-firestore>=2.15.0
google-cloud-storage==2.16.0
google-cloud-dataproc>=2.0.*